+++ src/main/java/com/side/cooktime/domain/member/service/GoogleService.java
@@ -1,4 +1,4 @@
-package com.side.cooktime.domain.member.controller;
+package com.side.cooktime.domain.member.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -6,28 +6,28 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.side.cooktime.domain.member.model.Member;
 import com.side.cooktime.domain.member.model.dto.GoogleUserInfo;
 import com.side.cooktime.domain.member.repository.MemberRepository;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
+import jakarta.transaction.Transactional;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
-import org.springframework.web.bind.annotation.*;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.client.RestTemplate;
 
 import java.net.URI;
-import java.util.Map;
 import java.util.Optional;
 
+@Transactional
+@RequiredArgsConstructor
+@Service
 @Log4j2
-@RestController
-@RequestMapping("/api/v1")
-public class GoogleLoginController {
+public class GoogleService {
 
     private final String GRANT_TYPE = "authorization_code";
     private final String REDIRECT_URI = "http://localhost/api/v1/callback";
@@ -42,34 +42,20 @@ public class GoogleLoginController {
 
     private final MemberRepository memberRepository;
 
-    public GoogleLoginController(MemberRepository memberRepository) {
-        this.memberRepository = memberRepository;
-    }
-
-    //임시
-    //https://accounts.google.com/o/oauth2/v2/auth?response_type=code&client_id=595159133596-ibjnkksdef8bumsndm9vjn0cd4790jtr.apps.googleusercontent.com&scope=email%20profile%20https://www.googleapis.com/auth/user.gender.read%20https://www.googleapis.com/auth/user.birthday.read&redirect_uri=http://localhost/api/v1/callback
-    @GetMapping(value = "/callback")
-    public String GoogleSignCallback(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        String code = request.getParameter("code");
-        log.info("## [REQUEST] code : {}", request.getParameter("code"));
-        return code;
-    }
-
-    @PostMapping(value = "/google-login")
-    public String googleLogin(@RequestBody Map<String, String> requestMap) throws Exception {
-        log.info("## [REQUEST] authorizationCode : {}", requestMap.get("authorizationCode"));
-        String accessToken = getGoogleAccessToken(requestMap.get("authorizationCode"));
-        GoogleUserInfo googleUserInfo = getUserInfo(accessToken);
+    public String login(String authorizationCode) throws JsonProcessingException {
+        log.info("## [REQUEST] authorizationCode : {}", authorizationCode);
+        String accessToken = getGoogleAccessToken(authorizationCode);
+        GoogleUserInfo googleUserInfo = getGoogleUserInfo(accessToken);
 
         Optional<Member> optionalMember = memberRepository.findByEmail(googleUserInfo.getEmail());
 
-        if(optionalMember.isEmpty()){
+        if (optionalMember.isEmpty()) {
             //TODO
             log.info("회원가입 진행 및 JWT Response");
             String provider = "google";
             String providerId = googleUserInfo.getId();
 
-        }else{
+        } else {
             //TODO
             log.info("JWT Response");
         }
@@ -77,8 +63,8 @@ public class GoogleLoginController {
         return googleUserInfo.getEmail();
     }
 
-    private GoogleUserInfo getUserInfo(String accessToken) {
-        URI uri = URI.create(USER_INFO_URL+"?access_token="+ accessToken);
+    private GoogleUserInfo getGoogleUserInfo(String accessToken) {
+        URI uri = URI.create(USER_INFO_URL + "?access_token=" + accessToken);
         RestTemplate restTemplate = new RestTemplate();
         GoogleUserInfo googleUserInfo = null;
         try {
@@ -132,6 +118,4 @@ public class GoogleLoginController {
         log.info("Access Token : {}", accessToken);
         return accessToken;
     }
-
-
 }
